backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
bandit
google-genai
